chore: rename navigationHelpers to useSubmissionContext
