+++ src/app/Scenes/SellWithArtsy/ArtworkForm/Components/SubmitArtworkFromMyCollection.tests.tsx
@@ -5,7 +5,7 @@ import { setupWithSubmitArtworkTestWrappers } from "app/Scenes/SellWithArtsy/Art
 
 const mockNavigateToNextStep = jest.fn()
 
-jest.mock("app/Scenes/SellWithArtsy/ArtworkForm/Utils/navigationHelpers", () => ({
+jest.mock("app/Scenes/SellWithArtsy/ArtworkForm/Utils/useSubmissionContext", () => ({
   useSubmissionContext: () => {
     return {
       navigateToNextStep: mockNavigateToNextStep,
